restructured code and introduced functors
