+++ kostyukov.mikhail/F0/freqDict.hpp
@@ -12,6 +12,7 @@ namespace kostyukov
     std::map< std::string, size_t > counts;
     size_t totalWords = 0;
   };
+
   using FreqDictManager = std::map< std::string, FrequencyDictionary >;
 
   struct WordFreqPair
@@ -20,136 +21,21 @@ namespace kostyukov
     double freq;
   };
 
-  struct MapToPairTransformer
-  {
-    MapToPairTransformer(size_t totalWords);
-    WordFreqPair operator()(const std::pair< const std::string, size_t >& mapPair) const;
-    private:
-      size_t totalWords_;
-  };
-
-  struct FreqComparator
-  {
-    explicit FreqComparator(bool ascending = false);
-    bool operator()(const WordFreqPair& first, const WordFreqPair& second) const;
-    private:
-      bool ascending_;
-  };
-
-  struct PairPrinter
-  {
-    explicit PairPrinter(std::ostream& out);
-    void operator()(const WordFreqPair&) const;
-    private:
-      std::ostream& out_;
-  };
-
-  struct KeyExtractor
-  {
-    std::string operator()(const std::pair< const std::string, size_t >& mapPair) const;
-  };
-
-  struct RemoveBatchPredicate
-  {
-    RemoveBatchPredicate(size_t total, bool less, double n);
-    bool operator()(const std::pair< const std::string, size_t >& pair) const;
-    private:
-      size_t totalWords_;
-      bool isLess_;
-      double threshold_;
-  };
-
-  struct DictDeleter
-  {
-    explicit DictDeleter(FrequencyDictionary& dict);
-    void operator()(const std::pair< const std::string, size_t >& pairToRemove) const;
-    private:
-      FrequencyDictionary& dict_;
-  };
-
-  struct SameWordPrinter
-  {
-    std::ostream& out;
-    const FrequencyDictionary& dict1;
-    const FrequencyDictionary& dict2;
-    const std::string& name1;
-    const std::string& name2;
-    void operator()(const std::string& key) const;
-  };
-
-  struct RangeFreqPredicate
-  {
-    RangeFreqPredicate(double minVal, double maxVal);
-    bool operator()(const WordFreqPair& pair) const;
-    private:
-      double minVal_;
-      double maxVal_;
-  };
-
-  struct DictNamePrinter
-  {
-    std::ostream& out;
-    void operator()(const std::pair< const std::string, FrequencyDictionary >& dictPair) const;
-  };
-
-  struct OutOfRangeFreqPredicate
-  {
-    OutOfRangeFreqPredicate(double minVal, double maxVal);
-    bool operator()(const WordFreqPair& pair) const;
-    private:
-      double minVal_;
-      double maxVal_;
-  };
-
-  struct MergeFunctor
-  {
-    explicit MergeFunctor(FrequencyDictionary& targetDict);
-    void operator()(const std::pair< const std::string, size_t >& pair);
-    private:
-      FrequencyDictionary& targetDict_;
-  };
-
-  struct IntersectFunctor
-  {
-    IntersectFunctor(FrequencyDictionary& targetDict, const FrequencyDictionary& otherDict, bool useMax);
-    void operator()(const std::pair< const std::string, size_t >& pair);
-    private:
-      FrequencyDictionary& targetDict_;
-      const FrequencyDictionary& otherDict_;
-      bool useMax_;
-  };
-
-  struct IsInvalidChar
-  {
-    bool operator()(char c) const;
-  };
-  struct IsPunct
-  {
-    bool operator()(const char c) const;
-  };
-  struct WordProcessor
-  {
-    explicit WordProcessor(FrequencyDictionary& dict);
-    void operator()(std::string word);
-    private:
-      FrequencyDictionary& dict_;
-  };
-
   void printHelp(std::ostream& out);
   void createDict(std::istream& in, std::ostream& out, FreqDictManager& dicts);
-  void top(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
-  void bottom(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
-  void deleteDict(std::istream& in, std::ostream& out, FreqDictManager& dicts);
   void loadDict(std::istream& in, std::ostream& out, FreqDictManager& dicts);
   void getFreq(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
-  void removeBatch(std::istream& in, std::ostream& out, FreqDictManager& dicts);
-  void findUniq(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
-  void findSame(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
-  void clear(std::istream& in, std::ostream& out, FreqDictManager& dicts);
+  void top(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
+  void bottom(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
   void listDicts(std::ostream& out, const FreqDictManager& dicts);
-  void rangeFreq(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
-  void outRangeFreq(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
+  void deleteDict(std::istream& in, std::ostream& out, FreqDictManager& dicts);
   void merge(std::istream& in, std::ostream& out, FreqDictManager& dicts);
   void intersect(std::istream& in, std::ostream& out, FreqDictManager& dicts);
+  void clear(std::istream& in, std::ostream& out, FreqDictManager& dicts);
+  void findUniq(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
+  void findSame(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
+  void removeBatch(std::istream& in, std::ostream& out, FreqDictManager& dicts);
+  void rangeFreq(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
+  void outRangeFreq(std::istream& in, std::ostream& out, const FreqDictManager& dicts);
 }
 #endif
